+++ src/links.rs
@@ -17,4 +17,6 @@ pub mod users {
     pub const BLOCKS_IDS: &'static str = "https://api.twitter.com/1.1/blocks/ids.json";
     pub const MUTES_LIST: &'static str = "https://api.twitter.com/1.1/mutes/users/list.json";
     pub const MUTES_IDS: &'static str = "https://api.twitter.com/1.1/mutes/users/ids.json";
+    pub const FOLLOW: &'static str = "https://api.twitter.com/1.1/friendships/create.json";
+    pub const UNFOLLOW: &'static str = "https://api.twitter.com/1.1/friendships/destroy.json";
 }