+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [Pending]
 ### Added
-- New functions to load the muted users list
+- New functions to load the muted users list (`mutes`, `mutes_ids`)
+- New functions to follow/unfollow a user (`follow`, `unfollow`)
 
 ### Changed
 - Combined IDLoader and UserLoader into CursorIter (with the same interface)